training/requirements.txt
tensorflow==2.15.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.37.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3